List contents of directory

+++ NMSSH/libssh2/libssh2_sftp.h
@@ -230,7 +230,7 @@ LIBSSH2_API LIBSSH2_SFTP_HANDLE *libssh2_sftp_open_ex(LIBSSH2_SFTP *sftp,
     libssh2_sftp_open_ex((sftp), (filename), (unsigned int)strlen(filename), (flags), \
                          (mode), LIBSSH2_SFTP_OPENFILE)
 #define libssh2_sftp_opendir(sftp, path) \
-    libssh2_sftp_open_ex((sftp), (path), strlen(path), 0, 0, \
+    libssh2_sftp_open_ex((sftp), (path), (unsigned int)strlen(path), 0, 0, \
                          LIBSSH2_SFTP_OPENDIR)
 
 LIBSSH2_API ssize_t libssh2_sftp_read(LIBSSH2_SFTP_HANDLE *handle,
@@ -242,7 +242,7 @@ LIBSSH2_API int libssh2_sftp_readdir_ex(LIBSSH2_SFTP_HANDLE *handle, \
                                         size_t longentry_maxlen,
                                         LIBSSH2_SFTP_ATTRIBUTES *attrs);
 #define libssh2_sftp_readdir(handle, buffer, buffer_maxlen, attrs)      \
-    libssh2_sftp_readdir_ex((handle), (buffer), (buffer_maxlen), NULL, 0, \
+    libssh2_sftp_readdir_ex((handle), (buffer), (unsigned long)(buffer_maxlen), NULL, 0, \
                             (attrs))
 
 LIBSSH2_API ssize_t libssh2_sftp_write(LIBSSH2_SFTP_HANDLE *handle,
